yaml_to_crontab_job_manager/requirements.txt
pyyaml>=6.0
python-crontab>=3.0
